Update the review structure
