+++ backend/src/services/review.ts
@@ -6,7 +6,7 @@ import authorisation from '../connectors/authorisation/index.js'
 import AccessRequestModel, { AccessRequestDoc } from '../models/AccessRequest.js'
 import ModelModel, { ModelDoc, ModelInterface } from '../models/Model.js'
 import ReleaseModel, { ReleaseDoc } from '../models/Release.js'
-import { Decision, DecisionKeys } from '../models/Response.js'
+import ResponseModel, { Decision, DecisionKeys, ResponseInterface, ResponseKind } from '../models/Response.js'
 import ReviewModel, { ReviewDoc, ReviewInterface } from '../models/Review.js'
 import ReviewRoleModel, { ReviewRoleDoc, ReviewRoleInterface } from '../models/ReviewRole.js'
 import { UserInterface } from '../models/User.js'
@@ -122,6 +122,38 @@ export async function findReviews(
   )
 
   return reviews.filter((_, i) => auths[i].success)
+}
+
+export async function getResponses(parentId) {
+  const [comments, reviews] = await Promise.all([
+    ResponseModel.find({ parentId, kind: ResponseKind.Comment }),
+    ReviewModel.aggregate<{ responses: ResponseInterface[] }>([
+      { $match: { parentId } },
+      {
+        $lookup: {
+          from: 'v2_responses',
+          let: { reviewId: '$_id' },
+          pipeline: [
+            {
+              $match: {
+                $expr: { $eq: ['$parentId', '$$reviewId'] },
+                kind: ResponseKind.Review,
+              },
+            },
+            { $sort: { createdAt: -1 } },
+          ],
+          as: 'responses',
+        },
+      },
+      { $project: { responses: 1, _id: 0 } },
+    ]),
+  ])
+
+  const responses = [...comments, ...reviews.flatMap(({ responses: reviewResponses }) => reviewResponses)].sort(
+    (first, second) => new Date(first.createdAt).getTime() - new Date(second.createdAt).getTime(),
+  )
+
+  return responses
 }
 
 export async function createReleaseReviews(model: ModelDoc, release: ReleaseDoc) {
@@ -376,7 +408,7 @@ export async function findReviewRoles(schemaIds?: string[]): Promise<ReviewRoleD
       throw BadReq('Unable to find schemas', { schemaIds })
     }
     const uniqueRoles = [...new Set(schemas.flatMap((s) => s.reviewRoles))]
-    mongoQuery.shortName = { $in: uniqueRoles, $ne: 'dro' }
+    mongoQuery.shortName = { $in: uniqueRoles }
   }
 
   return await ReviewRoleModel.find(mongoQuery)
